+++ packages/database/src/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './instrument-master';
+export * from './market-data';